RAG-LangChain-nvidia:llama-3.2-nv-embedqa-1b-v2/requirement.txt
langchain-community
langchain-nvidia-ai-endpoints
chromadb
numpy
python-dotenv
nltk
